refactor(ui): share Card, Callout, Toggle, and HelpDetails
- Promote duplicated settings chrome from CLI and extension options
  into packages/ui as general presentational primitives
- Point Settings/Options consumers at the shared exports
- Collapse CLI settings types/route helpers into the settings modules
