+++ apps/extension/src/options/Options.tsx
@@ -11,10 +11,18 @@ import { getAutoOpenOnFilesTab, setAutoOpenOnFilesTab } from "@extension/lib/pre
 import { getProviderSettings, setProviderSettings } from "@extension/lib/settings";
 import { requestTestConnection } from "@extension/lib/messaging";
 import { GitHubAuthSection } from "./GitHubAuthSection";
-import { SettingsCard } from "./SettingsCard";
-import { StatusCallout } from "./StatusCallout";
-import { Toggle } from "./Toggle";
-import { Button, cn, Input, Label, Select, Spinner, type SelectOption } from "@guided-review/ui";
+import {
+  Button,
+  Callout,
+  Card,
+  cn,
+  Input,
+  Label,
+  Select,
+  Spinner,
+  Toggle,
+  type SelectOption,
+} from "@guided-review/ui";
 
 interface ProviderIconProps {
   provider: ProviderId;
@@ -175,7 +183,7 @@ export function Options() {
       <div className="flex flex-col gap-4">
         <GitHubAuthSection />
 
-        <SettingsCard
+        <Card
           title="AI Provider"
           titleId="ai-provider-heading"
           description={
@@ -257,13 +265,11 @@ export function Options() {
               </Button>
             </div>
 
-            {statusMessage && (
-              <StatusCallout kind={statusMessage.kind} message={statusMessage.message} />
-            )}
+            {statusMessage && <Callout kind={statusMessage.kind} message={statusMessage.message} />}
           </div>
-        </SettingsCard>
+        </Card>
 
-        <SettingsCard
+        <Card
           title="Review"
           titleId="review-heading"
           description="How Guided Review behaves when you open a pull request."
@@ -289,7 +295,7 @@ export function Options() {
               aria-describedby="autoOpenOnFilesTab-hint"
             />
           </div>
-        </SettingsCard>
+        </Card>
       </div>
 
       <p className="mt-6 m-0 text-sm leading-relaxed text-muted">
