+++ packages/ui/src/components/Toggle.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@guided-review/ui";
+import { cn } from "../cn";
 
 export function Toggle({
   id,
