+++ packages/ui/src/index.ts
@@ -15,3 +15,8 @@ export { Label } from "./components/Label";
 
 export { Select } from "./components/Select";
 export type { SelectHandle, SelectOption } from "./components/Select";
+
+export { Toggle } from "./components/Toggle";
+export { Card } from "./components/Card";
+export { Callout } from "./components/Callout";
+export { HelpDetails } from "./components/HelpDetails";
